myproject/requirements.txt
Django==4.2
django-import-export==3.1.0